Refactored SubjectsImporter to reuse parts for other entities
