+++ src/groovy/dbnp/importer/SubjectsImporter.groovy
@@ -8,18 +8,8 @@ import grails.util.Holders
 /**
  * Defines the interface for an exporter
  */
-public class SubjectsImporter implements Importer {
+public class SubjectsImporter extends StudyTemplateEntityImporter<Subject> {
     def messageSource = Holders.grailsApplication.mainContext.getBean('messageSource')
-    
-    /**
-     * SecUser that is used for authorization
-     */
-    SecUser user
-    
-    /**
-     * List with validation errors
-     */
-    protected List<ImportValidationError> errors = []
     
     /**
      * Returns an identifier that describes this importer
@@ -34,39 +24,6 @@ public class SubjectsImporter implements Importer {
      */
     public boolean supportsType(String type) {
         type in [ "clinicaldata", "subjects" ]
-    }
-    
-    /**
-     * Returns a map of parameters that should be set for this importer
-     */
-    public List<ImporterParameter> getParameters() {
-        [
-            new ImporterParameter(name: 'study', label: 'Study', type: 'select', values: Study.giveWritableStudies(user)),
-            new ImporterParameter(name: 'template', label: 'Template', type: 'select', values: Template.findAllByEntity(Subject))
-        ]
-    }
-    
-    /**
-     * Returns a list of header options to match the headers against.
-     * Each item should have an id and a name. A description can be specified with extra information
-     * 
-     * @param parameters        Map with settings for the parameters specified in the getParameters method
-     * @return 
-     * @see getParameters()
-     */
-    public List getHeaderOptions(def parameters) {
-        // Create a list of domain fields and template fields to match against
-        def fields = getAllFields(parameters)
-         
-        // Return a proper format
-        fields.collect { [ id: it.name, name: it.name, description: it.comment ] }
-    }
-
-    /**
-     * Returns a list of validation errors
-     */
-    public List<ImportValidationError> getValidationErrors() {
-        [] + errors
     }
     
     /**
@@ -87,26 +44,10 @@ public class SubjectsImporter implements Importer {
      *          false if the validation on any of the object has failed
      */
     public boolean validateData(def data, def mapping, def parameters) {
-        errors = []
+        resetValidationErrors()
         
-        // First create a list of required fields and see whether they are mapped.
-        // If not, not a single entity will be imported
-        def fields = getAllFields(parameters)
-        def mappingValues = mapping.values()
-        def notMappedRequiredFields = fields.findAll { it.required }.findAll { field ->
-            // Filter on required fields that have not been mapped
-            !mappingValues.find { fieldMapping -> fieldMapping?.field?.id == field.name }
-        }
-        
-        if( notMappedRequiredFields ) {
-            notMappedRequiredFields.each { 
-                errors << new ImportValidationError(
-                    code: 1,
-                    message: "The field " + it.name + " is a required field in this template, and must be mapped."
-                )
-            }
-            
-            return false;
+        if( !super.validateData(data, mapping, parameters) ) {
+            return false
         }
         
         // After that, validate each line. The header line is not needed anymore
@@ -138,8 +79,7 @@ public class SubjectsImporter implements Importer {
      *          false if the validation on any of the object has failed 
      */
     public boolean importData(def data, def mapping, def parameters) {
-        // Reset validation errors
-        errors = []
+        resetValidationErrors()
         
         // Now loop through each line and try to import the object.
         for( def lineNr = 1; lineNr < data.size(); lineNr++) {
@@ -164,73 +104,16 @@ public class SubjectsImporter implements Importer {
     }
     
     /**
-     * Creates an object, based on the specified parameters
-     * @param   data            List with the data from one line in the file, used to create this specific object 
-     * @param   mapping         Mapping from field number to object property name. The key in this map is the column number,
-     *                                  the value is a map that contains 2 entries: 
-     *                                          ignore (boolean)        Whether this column should be ignored
-     *                                          field (map)             Map describing the field selected. The format 
-     *                                                                  is the same as the output from getHeaderOptions 
-     * @param   parameters      Refers to a map with parameter values for the parameters needed by the importer
-     * @return  True if all objects were imported succesfully,
-     *          false if the validation on any of the object has failed
-     * @see     getHeaderOptions()
+     * Returns a new instance of the correct type
      */
-    public def createObject(def data, def mapping, def parameters) {
-        // Create an initial object
-        def object = new Subject(template: getTemplate(parameters))
-        
-        // Associate the object with its parent
-        object.parent = Study.get(parameters.study)
-        
-        // Loop through all columns
-        data.eachWithIndex { cell, columnIndex ->
-            // Retrieve the mapping
-            def columnMapping = mapping[columnIndex.toString()]
-
-            if( !columnMapping || columnMapping.ignore || !columnMapping.field?.id ) {
-                log.debug( "Ignoring column " + columnIndex )
-                return
-            }
-            
-            // Determine where to store this value
-            def fieldName = columnMapping.field.id
-            log.debug( "Setting column " + columnIndex + " to field " + fieldName )
-            
-            // Store the value itself
-            // TODO: Format and/or parse the value
-            object.setFieldValue(fieldName, cell, true)
-        }
-        
-        object
-    }
-
-    /**
-     * Retrieves a template selected by the user
-     */
-    protected Template getTemplate(parameters) {
-        def templateId = parameters?.template?.isLong() ? parameters.template.toLong() : null
-        def template
-        
-        // Load the template from the database
-        if( templateId ) {
-            template = Template.get(templateId)
-        }
-        
-        if( !template ) {
-            throw new IllegalArgumentException( "No template with the templateId " + templateId + " could be found." )
-        }
-        
-        template
+    public Subject newInstance(def parameters) {
+        new Subject(parameters)
     }
     
     /**
-     * Retrieves a list of domain and template fields, given the set of parameters
+     * Returns a list of domain fields for the current template entity
      */
-    protected List getAllFields(parameters) {
-        def template = getTemplate(parameters)
-        
-        // Create a list of domain fields and template fields to match against
-        Subject.domainFields + ( template.fields ?: [] )
+    public List getDomainFields() {
+        Subject.domainFields
     }
 }
